deployment/cdk/requirements.txt
aws-cdk-lib==2.144.0
constructs>=10.0.0,<11.0.0
cdk-ecr-deployment==3.0.127
boto3
gradio